Update OOP Solution.py
Fixed error where hit_positions wasn't being updated. The problem was that I was trying to access it as a class variable when it's an instance variable. Fixed the shoot and sunk Board methods to return the instance of the ship hit rather than the name and updated the instance's hit_positions appropriately.

Still an error in drawing, not sure why hits are being marked as O's for the submarine and Patrol ship. Not sure what's wrong